Reorganize src files for handlers, templates, etc.
