+++ server/routes.go
@@ -1,58 +1,12 @@
 package server
 
 import (
-	"encoding/json"
-	"html/template"
 	"net/http"
-	"path/filepath"
-	"strings"
 
 	"github.com/gorilla/mux"
 	"github.com/justinas/alice"
-	"github.com/pereztr5/cyboard/server/models"
 	"github.com/urfave/negroni"
 )
-
-type Page struct {
-	Title string
-	T     models.Team
-}
-
-var templates map[string]*template.Template
-
-// Parse templates at startup
-func ensureAppTemplates() {
-	if templates != nil {
-		return
-	}
-
-	templates = make(map[string]*template.Template)
-	funcMap := template.FuncMap{
-		"title":              strings.Title,
-		"totalChallenges":    getChallenges,
-		"teamChallenges":     getTeamChallenges,
-		"teamScore":          getTeamScore,
-		"allTeamScores":      getAllTeamScores,
-		"allBlueTeams":       DataGetTeams,
-		"getOwnedChalGroups": getChallengesOwnerOf,
-		"getStatus":          DataGetResultByService,
-		"serviceList":        DataGetServiceList,
-		"challengesList":     DataGetChallengeGroupsList,
-		"allUsers":           DataGetAllUsers,
-		"isChallengeOwner":   AllowedToConfigureChallenges,
-		"existsSpecialFlags": existsSpecialFlags,
-		"StringsJoin":        strings.Join,
-	}
-
-	includes := mustGlobFiles("tmpl/includes/*.tmpl")
-	layouts := mustGlobFiles("tmpl/*.tmpl")
-
-	for _, layout := range layouts {
-		files := append(includes, layout)
-		title := strings.TrimSuffix(filepath.Base(layout), ".tmpl")
-		templates[title] = template.Must(template.New(layout).Funcs(funcMap).ParseFiles(files...))
-	}
-}
 
 type methodHandlers map[string]http.HandlerFunc
 
@@ -145,185 +99,4 @@ func CreateWebRouter(teamScoreUpdater, servicesUpdater *broadcastHub) *mux.Route
 	})
 
 	return router
-}
-
-func ShowHome(w http.ResponseWriter, r *http.Request) {
-	t := r.Context().Value("team")
-	p := Page{Title: "homepage"}
-	if t != nil {
-		p.T = t.(models.Team)
-	}
-	renderTemplate(w, p)
-}
-
-func ShowLogin(w http.ResponseWriter, r *http.Request) {
-	if r.Context().Value("team") == nil {
-		p := Page{
-			Title: "login",
-		}
-		renderTemplate(w, p)
-	} else {
-		http.Redirect(w, r, "/dashboard", 302)
-	}
-}
-
-func SubmitLogin(w http.ResponseWriter, r *http.Request) {
-	session, err := Store.Get(r, "cyboard")
-	//if err != nil {
-	//	Logger.Warn("Getting session cookie from Store failed: ", err)
-	//}
-
-	succ := CheckCreds(w, r)
-	if succ {
-		err = session.Save(r, w)
-		if err != nil {
-			http.Error(w, http.StatusText(500), 500)
-			return
-		}
-		http.Redirect(w, r, "/dashboard", 302)
-		return
-	}
-	http.Redirect(w, r, "/login", 302)
-}
-
-func Logout(w http.ResponseWriter, r *http.Request) {
-	session, err := Store.Get(r, "cyboard")
-	if err != nil {
-		http.Error(w, http.StatusText(400), 400)
-		return
-	}
-
-	delete(session.Values, "id")
-	// Make sure we save the session after deleting the ID.
-	err = session.Save(r, w)
-	if err != nil {
-		http.Error(w, http.StatusText(500), 500)
-		return
-	}
-
-	http.Redirect(w, r, "/login", 302)
-}
-
-func ShowTeamDashboard(w http.ResponseWriter, r *http.Request) {
-	p := Page{
-		Title: "dashboard",
-		T:     r.Context().Value("team").(models.Team),
-	}
-	renderTemplate(w, p)
-}
-
-func ShowChallenges(w http.ResponseWriter, r *http.Request) {
-	t := r.Context().Value("team")
-	if t != nil {
-		p := Page{
-			Title: "challenges",
-			T:     t.(models.Team),
-		}
-		renderTemplate(w, p)
-	}
-}
-
-func ShowScoreboard(w http.ResponseWriter, r *http.Request) {
-	t := r.Context().Value("team")
-	p := Page{Title: "scoreboard"}
-	if t != nil {
-		p.T = t.(models.Team)
-	}
-	renderTemplate(w, p)
-}
-
-func ShowServices(w http.ResponseWriter, r *http.Request) {
-	t := r.Context().Value("team")
-	p := Page{Title: "services"}
-	if t != nil {
-		p.T = t.(models.Team)
-	}
-	renderTemplate(w, p)
-}
-
-func GetScores(w http.ResponseWriter, r *http.Request) {
-	scores := DataGetAllScore()
-	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-	if err := json.NewEncoder(w).Encode(scores); err != nil {
-		Logger.Error("Error encoding json: ", err)
-	}
-}
-
-func GetScoresSplit(w http.ResponseWriter, r *http.Request) {
-	scores := DataGetAllScoreSplitByType()
-	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-	if err := json.NewEncoder(w).Encode(scores); err != nil {
-		Logger.Error("Error encoding json: ", err)
-	}
-}
-
-func GetServices(w http.ResponseWriter, r *http.Request) {
-	services := DataGetServiceList()
-	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-	if err := json.NewEncoder(w).Encode(services); err != nil {
-		Logger.Error("Error encoding json: ", err)
-	}
-}
-
-func getChallenges() []ChallengeCount {
-	totals, err := DataGetTotalChallenges()
-	if err != nil {
-		Logger.Error("Could not get challenges: ", err)
-	}
-	return totals
-}
-
-func getTeamChallenges(teamname string) []ChallengeCount {
-	acquired, err := DataGetTeamChallenges(teamname)
-	if err != nil {
-		Logger.Error("Could not get team challenges: ", err)
-	}
-	return acquired
-}
-
-func getChallengesOwnerOf(adminof, teamgroup string) []string {
-	switch teamgroup {
-	case "admin", "blackteam":
-		return DataGetChallengeGroupsList()
-	default:
-		return []string{adminof}
-	}
-}
-
-func getTeamScore(teamname string) int {
-	return DataGetTeamScore(teamname)
-}
-
-func getAllTeamScores() []map[string]interface{} {
-	results := DataGetAllScoreSplitByType()
-	scores := make([]map[string]interface{}, 0, len(results)/2)
-
-	acc := make(map[string]map[string]interface{})
-	for _, r := range results {
-		score, ok := acc[r.Teamname]
-		if ok {
-			score[r.Type] = r.Points
-			score["Points"] = score["CTF"].(int) + score["Service"].(int)
-			scores = append(scores, score)
-		} else {
-			acc[r.Teamname] = map[string]interface{}{
-				"Teamnumber": r.Teamnumber,
-				"Teamname":   r.Teamname,
-				r.Type:       r.Points,
-			}
-		}
-	}
-
-	return scores
-}
-
-func existsSpecialFlags() bool {
-	return len(specialChallenges) > 0
-}
-
-func renderTemplate(w http.ResponseWriter, p Page) {
-	err := templates[p.Title].ExecuteTemplate(w, p.Title+".tmpl", &p)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-	}
 }
